feat(frontend): extract tier list to sub component
